# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{projectId}/regions": {
      "post": {
        "operationId": "RegionController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Región creada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRegionDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos."
          },
          "404": {
            "description": "Región padre no encontrada."
          },
          "409": {
            "description": "languageCode ya existe."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear una nueva región para un proyecto específico",
        "tags": [
          "Regions"
        ]
      },
      "get": {
        "operationId": "RegionController_findAll",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de regiones.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateRegionDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener todas las regiones para un proyecto específico",
        "tags": [
          "Regions"
        ]
      }
    },
    "/api/projects/{projectId}/regions/{languageCode}": {
      "get": {
        "operationId": "RegionController_findOne",
        "parameters": [
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Código de idioma (ID) de la región",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Región encontrada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRegionDto"
                }
              }
            }
          },
          "404": {
            "description": "Proyecto o Región no encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener una región específica dentro de un proyecto",
        "tags": [
          "Regions"
        ]
      },
      "patch": {
        "operationId": "RegionController_update",
        "parameters": [
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Código de idioma (ID) de la región a actualizar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRegionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Región actualizada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRegionDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos (languageCode no se puede cambiar)."
          },
          "404": {
            "description": "Proyecto o Región no encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar una región específica dentro de un proyecto",
        "tags": [
          "Regions"
        ]
      },
      "delete": {
        "operationId": "RegionController_remove",
        "parameters": [
          {
            "name": "languageCode",
            "required": true,
            "in": "path",
            "description": "Código de idioma (ID) de la región a eliminar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Región eliminada."
          },
          "404": {
            "description": "Proyecto o Región no encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar una región específica dentro de un proyecto",
        "tags": [
          "Regions"
        ]
      }
    },
...
    "/api/projects/{projectId}/rag-document-metadata": {
      "post": {
        "operationId": "RagDocumentMetadataController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRagDocumentMetadataDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Metadatos creados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RagDocumentMetadataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos."
          },
          "404": {
            "description": "Proyecto o Región referenciada no encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear metadatos para un documento RAG dentro de un proyecto",
        "tags": [
          "RAG Document Metadata"
        ]
      },
      "get": {
        "operationId": "RagDocumentMetadataController_findAll",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de metadatos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RagDocumentMetadataResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Proyecto no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener todos los metadatos de documentos RAG de un proyecto",
        "tags": [
          "RAG Document Metadata"
        ]
      }
    },
    "/api/projects/{projectId}/rag-document-metadata/{metadataId}": {
      "get": {
        "operationId": "RagDocumentMetadataController_findOne",
        "parameters": [
          {
            "name": "metadataId",
            "required": true,
            "in": "path",
            "description": "ID de los metadatos (CUID)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metadatos encontrados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RagDocumentMetadataResponse"
                }
              }
            }
          },
          "404": {
            "description": "Proyecto o Metadatos no encontrados."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener metadatos por ID dentro de un proyecto",
        "tags": [
          "RAG Document Metadata"
        ]
      },
      "patch": {
        "operationId": "RagDocumentMetadataController_update",
        "parameters": [
          {
            "name": "metadataId",
            "required": true,
            "in": "path",
            "description": "ID a actualizar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRagDocumentMetadataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Metadatos actualizados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RagDocumentMetadataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos."
          },
          "404": {
            "description": "Proyecto, Metadatos o Región referenciada no encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar metadatos por ID dentro de un proyecto",
        "tags": [
          "RAG Document Metadata"
        ]
      },
      "delete": {
        "operationId": "RagDocumentMetadataController_remove",
        "parameters": [
          {
            "name": "metadataId",
            "required": true,
            "in": "path",
            "description": "ID a eliminar",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID del proyecto",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Metadatos eliminados."
          },
          "404": {
            "description": "Proyecto o Metadatos no encontrados."
          },
          "409": {
            "description": "Conflicto al eliminar (referenciado por otras entidades)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar metadatos por ID dentro de un proyecto",
        "tags": [
          "RAG Document Metadata"
        ]
      }
    },
...
  },
  "info": {
    "title": "JAPM API",
    "description": "API para la aplicación de Prompt Engineering JAPM",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateRegionDto": {
        "type": "object",
        "properties": {
          "languageCode": {
            "type": "string",
            "minLength": 5,
            "maxLength": 5,
            "description": "Código de idioma único que actúa como ID",
            "example": "de-DE"
          },
          "name": {
            "type": "string",
            "description": "Nombre de la región",
            "example": "Alemania"
          },
          "parentRegionId": {
            "type": "string",
            "description": "languageCode de la región padre (opcional)",
            "example": "eu"
          },
          "timeZone": {
            "type": "string",
            "description": "Zona horaria",
            "example": "Europe/Berlin"
          },
          "defaultFormalityLevel": {
            "type": "string",
            "description": "Nivel de formalidad por defecto (opcional)",
            "example": "Formal"
          },
          "notes": {
            "type": "string",
            "description": "Notas adicionales (opcional)"
          }
        },
        "required": [
          "languageCode",
          "name"
        ]
      },
      "UpdateRegionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre de la región",
            "example": "Alemania"
          },
          "parentRegionId": {
            "type": "string",
            "description": "languageCode de la región padre (opcional)",
            "example": "eu"
          },
          "timeZone": {
            "type": "string",
            "description": "Zona horaria",
            "example": "Europe/Berlin"
          },
          "defaultFormalityLevel": {
            "type": "string",
            "description": "Nivel de formalidad por defecto (opcional)",
            "example": "Formal"
          },
          "notes": {
            "type": "string",
            "description": "Notas adicionales (opcional)"
          }
        }
      },
...
      "CreateRagDocumentMetadataDto": {
        "type": "object",
        "properties": {
          "regionId": {
            "type": "string",
            "format": "uuid",
            "description": "ID de la región asociada (opcional)"
          },
          "documentName": {
            "type": "string",
            "description": "Nombre del documento RAG",
            "example": "policy_update_q1_2024.pdf"
          },
          "category": {
            "type": "string",
            "description": "Categoría (opcional)",
            "example": "Compliance"
          },
          "complianceReviewed": {
            "type": "boolean",
            "description": "Indica si ha sido revisado por compliance",
            "default": false
          },
          "piiRiskLevel": {
            "type": "string",
            "description": "Nivel de riesgo PII (Información Personal Identificable) (opcional)",
            "example": "Low"
          },
          "lastReviewedBy": {
            "type": "string",
            "description": "Identificador de quién lo revisó por última vez (opcional)",
            "example": "user_cuid_123"
          }
        },
        "required": [
          "documentName"
        ]
      },
      "RagDocumentMetadataResponse": {
        "type": "object",
        "properties": {
          "regionId": {
            "type": "string",
            "format": "uuid",
            "description": "ID de la región asociada (opcional)"
          },
          "documentName": {
            "type": "string",
            "description": "Nombre del documento RAG",
            "example": "policy_update_q1_2024.pdf"
          },
          "category": {
            "type": "string",
            "description": "Categoría (opcional)",
            "example": "Compliance"
          },
          "complianceReviewed": {
            "type": "boolean",
            "description": "Indica si ha sido revisado por compliance",
            "default": false
          },
          "piiRiskLevel": {
            "type": "string",
            "description": "Nivel de riesgo PII (Información Personal Identificable) (opcional)",
            "example": "Low"
          },
          "lastReviewedBy": {
            "type": "string",
            "description": "Identificador de quién lo revisó por última vez (opcional)",
            "example": "user_cuid_123"
          },
          "region": {
            "$ref": "#/components/schemas/CreateRegionDto"
          },
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          }
        },
        "required": [
          "documentName",
          "id",
          "projectId"
        ]
      },
      "UpdateRagDocumentMetadataDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}